acerto application

+++ src/main/java/mprj/mp/br/calculos/domain/jpa/IGPM.java
@@ -1,16 +1,9 @@
 package mprj.mp.br.calculos.domain.jpa;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.format.annotation.DateTimeFormat;
-
-import java.io.Serializable;
 
 import javax.persistence.*;
-import java.io.Serializable;
-import java.util.ArrayList;
 import java.util.Date;
-import java.util.List;
 
 @Entity
 @Table(name = "tbl_igpm")
@@ -24,7 +17,7 @@ public class IGPM {
     private String Nome;
 
     @Column(name = "fator")
-    private double fator;
+    private float fator;
 
     @Column(name = "valor")
     private double valor;
@@ -43,8 +36,8 @@ public class IGPM {
     public void setValor(double valor){ this.valor = valor;}
     public Date getData() {return data;}
     public void  setData(Date data){ this.data = data;}
-    public double getFator(){return fator;}
-    public void setFator(double fator) { this.fator = fator;}
+    public float getFator(){return fator;}
+    public void setFator(float fator) { this.fator = fator;}
     public String getNome() { return Nome;}
     public void setNome(String Nome){this.Nome = Nome;}
     /*
